requirements.txt
flask
flask-cors
flask-limiter
python-dotenv
openai>=1.0.0
